+++ config/maieutic.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for Indra/LaravelMaieutic
+return [
+
+];