+++ src/LaravelMaieutic.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Indra\LaravelMaieutic;
+
+class LaravelMaieutic {}